+++ template/__init__.py
@@ -1,0 +1,2 @@
+from .tempita import include
+from .wrapper import gen_wrapper, get_clean_docstring
