+++ template/__main__.py
@@ -1,0 +1,4 @@
+from template.tempita import build
+
+if __name__ == '__main__':
+    build()
