+++ template/src/_array.pyx_
@@ -1,4 +1,3 @@
-from typing import List
 import numpy
 cimport numpy
 from numpy import ndarray
@@ -8,11 +7,11 @@ from copy import copy, deepcopy
 
 from ._value cimport BaseMutableTag
 from ._const cimport CommaSpace, ID_BYTE_ARRAY, ID_INT_ARRAY, ID_LONG_ARRAY
-from ._util cimport write_array, BufferContext, read_int, read_data, read_string
+from ._util cimport write_array, BufferContext, read_int, read_data
 from ._list cimport ListTag
 {{py:
 import numpy
-from tools.tempita import include, gen_wrapper, get_clean_docstring
+from template import include, gen_wrapper, get_clean_docstring
 }}
 
 
@@ -92,7 +91,7 @@ cdef class BaseArrayTag(BaseMutableTag):
         "view"
     ]
 )}}
-{{include("amulet_nbt/BaseMutableTag.pyx.in", cls_name="BaseArrayTag")}}
+{{include("BaseMutableTag.pyx.in", cls_name="BaseArrayTag")}}
     @property
     def shape(self):
 {{get_clean_docstring(numpy.ndarray.shape)}}
