+++ template/templates/BaseIntTag.pyx.in
@@ -1,4 +1,4 @@
-{{py:from tools.tempita import include, gen_wrapper}}
+{{py:from template import include, gen_wrapper}}
 {{gen_wrapper(
     "value_",
     int,
@@ -21,7 +21,7 @@
         return self.value_.conjugate()
     conjugate.__doc__ = int.conjugate.__doc__
 
-{{include("amulet_nbt/BaseNumericTag.pyx.in", cls_name=cls_name)}}
+{{include("BaseNumericTag.pyx.in", cls_name=cls_name)}}
 
     @classmethod
     def from_bytes(
