+++ template/templates/BaseMutableTag.pyx.in
@@ -1,5 +1,5 @@
-{{py:from tools.tempita import include}}
-{{include("amulet_nbt/BaseTag.pyx.in", cls_name=cls_name)}}
+{{py:from template import include}}
+{{include("BaseTag.pyx.in", cls_name=cls_name)}}
 
     @property
     def py_data({{cls_name}} self):
