+++ template/templates/BaseTag.pyx.in
